requirements.txt
asgiref==3.3.0
Django==3.1.3
django-js-asset==1.2.2
django-mptt==0.11.0
Pillow==8.0.1
pkg-resources==0.0.0
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
